[REF] pos*: unify owl app instantiation
pos*: point_of_sale, pos_self_order, pos_customer_display

There are multiple owl apps in the pos ecosystem.
For each app we follow a similar pattern of instantiation.
This means that code is repeated, which has often lead to
a worse developer experience.

In this commit we create a common function that performs
all the steps needed when starting an owl app and we adapt
the codebase to use it.
