+++ addons/pos_self_order/__manifest__.py
@@ -51,12 +51,12 @@
             "web/static/src/libs/fontawesome/css/font-awesome.css",
             "web/static/lib/odoo_ui_icons/*",
             'web/static/src/legacy/scss/ui.scss',
-            "point_of_sale/static/src/utils.js",
             'bus/static/src/services/bus_service.js',
             'bus/static/src/bus_parameters_service.js',
             'bus/static/src/multi_tab_service.js',
             'bus/static/src/workers/*',
             'barcodes/static/src/barcode_service.js',
+            'point_of_sale/static/src/startOwl.js',
             'point_of_sale/static/src/utils.js',
             'web/static/lib/bootstrap/js/dist/dom/data.js',
             'web/static/lib/bootstrap/js/dist/dom/event-handler.js',
@@ -66,7 +66,6 @@
             "web/static/lib/bootstrap/js/dist/carousel.js",
             'web/static/lib/bootstrap/js/dist/scrollspy.js',
             "point_of_sale/static/src/app/store/models/product_custom_attribute.js",
-            "point_of_sale/static/src/utils.js",
             'web_editor/static/src/js/editor/odoo-editor/src/base_style.scss',
             'web_editor/static/src/scss/web_editor.common.scss',
             "point_of_sale/static/src/app/generic_components/**/*",
